Initial Commit of the masking code.
This is the initial commit of the credit card masking code. Basic character length check. Eventually the code will check for a couple other conditions reflective of min length for Discover and Amex card types. Additional planned enhancements is to loop through excel doc and mask an entire of column of card numbers then save the doc locally on the users machine.